control/requirements.txt
pandas==1.3.4
tabula-py==2.4.0
unoconv==0.9.0
pypdf2==2.10.3
openpyxl==2.4
fuzzywuzzy==0.18.0